requirements.txt
requests
bs4
transformers
huggingface_hub
bitsandbytes 
peft
accelerate
datasets 
tqdm
streamlit
pyyaml